# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise College ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed

  # USERS
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/bulk:
    post:
      operationId: bulkMarkAttendance
      tags: [attendance]
      summary: Bulk mark attendance for a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAttendanceInput"
      responses:
        "201":
          description: Attendance marked
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Get attendance summary by student and course
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
        - name: semesterId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceSummary"

  # EXAMINATIONS
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: facultyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assignments/{id}/submit:
    post:
      operationId: submitAssignment
      tags: [assignments]
      summary: Submit assignment (student)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /assignments/{id}/submissions:
    get:
      operationId: listSubmissions
      tags: [assignments]
      summary: List submissions for assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"

  # LIBRARY
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string

    # USER SCHEMAS
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    AttendanceRecord:
      type: object
      required: [id, studentId, courseId, date, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        date:
          type: string
        status:
          type: string
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    AttendanceInput:
      type: object
      required: [studentId, courseId, date, status]
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        date:
          type: string
        status:
          type: string
        remarks:
          type: string
    BulkAttendanceInput:
      type: object
      required: [courseId, date, records]
      properties:
        courseId:
          type: integer
        date:
          type: string
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceBulkItem"
    AttendanceBulkItem:
      type: object
      required: [studentId, status]
      properties:
        studentId:
          type: integer
        status:
          type: string
        remarks:
          type: string
    AttendanceSummary:
      type: object
      required: [studentId, studentName, rollNumber, courseId, courseName, totalClasses, presentCount, absentCount, percentage]
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        rollNumber:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        totalClasses:
          type: integer
        presentCount:
          type: integer
        absentCount:
          type: integer
        percentage:
          type: number

    # EXAMINATION SCHEMAS
...
    Assignment:
      type: object
      required: [id, title, courseId, facultyId, dueDate, totalMarks, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        facultyId:
          type: integer
        facultyName:
          type: ["string", "null"]
        dueDate:
          type: string
        totalMarks:
          type: integer
        submissionCount:
          type: ["integer", "null"]
        status:
          type: string
        createdAt:
          type: string
    AssignmentInput:
      type: object
      required: [title, courseId, facultyId, dueDate, totalMarks]
      properties:
        title:
          type: string
        description:
          type: string
        courseId:
          type: integer
        facultyId:
          type: integer
        dueDate:
          type: string
        totalMarks:
          type: integer
    AssignmentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        courseId:
          type: integer
        dueDate:
          type: string
        totalMarks:
          type: integer
        status:
          type: string
    Submission:
      type: object
      required: [id, assignmentId, studentId, submittedAt, createdAt]
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        marksObtained:
          type: ["number", "null"]
        feedback:
          type: ["string", "null"]
        status:
          type: string
        submittedAt:
          type: string
        createdAt:
          type: string
    SubmissionInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
        content:
          type: string
        fileUrl:
          type: string

    # LIBRARY SCHEMAS
...